fix(Base16384): 替换已弃用函数
- 在 encodeToString 函数中，将字节转换为字符时使用 toInt().toChar() 方法

+++ src/main/kotlin/Base16384.kt
@@ -101,7 +101,7 @@ object Base16384 {
 
     fun encodeToString(input: String): String {
         val encodedArray = encode(input)
-        return encodedArray.joinToString("") { it.toChar().toString() }
+        return encodedArray.joinToString("") { it.toInt().toChar().toString() }
     }
 
     fun decodeFromString(input: String): String {
